// src/app/core/models/product.ts
export class Product {
  $key: string;
  productId: string;
  productName: string;
  productCategory: string;
  productPrice: string;
  productCost: string;
  productUM: string;
  productAmount: string;
  productDescription: string;
  productImageUrl: string;
  productProvince: string;
  productProvinces: string[];
  productSeller: string;
  productAgency: string;
  productAgencys: string[];
  productState: boolean;
  products: Product[];
}